+++ simple_history/middleware.py
@@ -1,7 +1,7 @@
-from . models import HistoricalRecords
+from .models import HistoricalRecords
 
 
-class HistoryRequestMiddleware(object):
+class HistoryRequestMiddleware:
     """Expose request to HistoricalRecords.
 
     This middleware sets request as a local thread variable, making it
@@ -9,5 +9,21 @@ class HistoryRequestMiddleware(object):
     authenticated user making a change.
     """
 
-    def process_request(self, request):
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        # Code to be executed for each request before
+        # the view (and later middleware) are called.
+
         HistoricalRecords.thread.request = request
+
+        response = self.get_response(request)
+
+        # Code to be executed for each request/response after
+        # the view is called.
+
+        if hasattr(HistoricalRecords.thread, 'request'):
+            del HistoricalRecords.thread.request
+
+        return response