+++ simple_history/models.py
@@ -1,40 +1,24 @@
 from __future__ import unicode_literals
 
-import threading
 import copy
-import warnings
+import importlib
+import threading
+import uuid
 
-import django
-from django.db import models, router
-from django.db.models.fields.proxy import OrderWrt
-from django.db.models.fields.related import RelatedField
+from django.apps import apps
 from django.conf import settings
 from django.contrib import admin
+from django.db import models, router
+from django.db.models import Q
+from django.db.models.fields.proxy import OrderWrt
+from django.urls import reverse
 from django.utils import six
-from django.utils.encoding import python_2_unicode_compatible
-from django.utils.encoding import smart_text
+from django.utils.encoding import python_2_unicode_compatible, smart_text
+from django.utils.text import format_lazy
 from django.utils.timezone import now
-from django.utils.translation import string_concat
+from django.utils.translation import ugettext_lazy as _
 
-try:
-    import importlib
-except ImportError:
-    from django.utils import importlib
-try:
-    from django.apps import apps
-except ImportError:  # Django < 1.7
-    from django.db.models.loading import get_app
-    apps = None
-else:
-    get_app = apps.get_app
-try:
-    from south.modelsinspector import add_introspection_rules
-except ImportError:  # south not present
-    pass
-else:  # south configuration for CustomForeignKeyField
-    add_introspection_rules(
-        [], ["^simple_history.models.CustomForeignKeyField"])
-
+from . import exceptions
 from .manager import HistoryDescriptor
 
 registered_models = {}
@@ -44,9 +28,17 @@ class HistoricalRecords(object):
     thread = threading.local()
 
     def __init__(self, verbose_name=None, bases=(models.Model,),
-                 user_related_name='+'):
+                 user_related_name='+', table_name=None, inherit=False,
+                 history_id_field=None,
+                 excluded_fields=None):
         self.user_set_verbose_name = verbose_name
         self.user_related_name = user_related_name
+        self.table_name = table_name
+        self.inherit = inherit
+        self.history_id_field = history_id_field
+        if excluded_fields is None:
+            excluded_fields = []
+        self.excluded_fields = excluded_fields
         try:
             if isinstance(bases, six.string_types):
                 raise TypeError
@@ -57,7 +49,8 @@ class HistoricalRecords(object):
     def contribute_to_class(self, cls, name):
         self.manager_name = name
         self.module = cls.__module__
-        models.signals.class_prepared.connect(self.finalize, sender=cls)
+        self.cls = cls
+        models.signals.class_prepared.connect(self.finalize, weak=False)
         self.add_extra_methods(cls)
 
     def add_extra_methods(self, cls):
@@ -73,12 +66,34 @@ class HistoricalRecords(object):
             finally:
                 del self.skip_history_when_saving
             return ret
+
         setattr(cls, 'save_without_historical_record',
                 save_without_historical_record)
 
     def finalize(self, sender, **kwargs):
-        history_model = self.create_history_model(sender)
-        module = importlib.import_module(self.module)
+        inherited = False
+        try:
+            hint_class = self.cls
+        except AttributeError:  # called via `register`
+            pass
+        else:
+            if hint_class is not sender:  # set in concrete
+                inherited = (self.inherit and issubclass(sender, hint_class))
+                if not inherited:
+                    return  # set in abstract
+        if hasattr(sender._meta, 'simple_history_manager_attribute'):
+            raise exceptions.MultipleRegistrationsError(
+                '{}.{} registered multiple times for history tracking.'.format(
+                    sender._meta.app_label,
+                    sender._meta.object_name,
+                )
+            )
+        history_model = self.create_history_model(sender, inherited)
+        if inherited:
+            # Make sure history model is in same module as concrete model
+            module = importlib.import_module(history_model.__module__)
+        else:
+            module = importlib.import_module(self.module)
         setattr(module, history_model.__name__, history_model)
 
         # The HistoricalRecords object will be discarded,
@@ -92,35 +107,44 @@ class HistoricalRecords(object):
         setattr(sender, self.manager_name, descriptor)
         sender._meta.simple_history_manager_attribute = self.manager_name
 
-    def create_history_model(self, model):
+    def create_history_model(self, model, inherited):
         """
         Creates a historical model to associate with the model provided.
         """
         attrs = {'__module__': self.module}
 
         app_module = '%s.models' % model._meta.app_label
-        if model.__module__ != self.module:
+
+        if inherited:
+            # inherited use models module
+            attrs['__module__'] = model.__module__
+        elif model.__module__ != self.module:
             # registered under different app
             attrs['__module__'] = self.module
         elif app_module != self.module:
-            if apps is None:  # Django < 1.7
-                # has meta options with app_label
-                app = get_app(model._meta.app_label)
-                attrs['__module__'] = app.__name__  # full dotted name
-            else:
-                # Abuse an internal API because the app registry is loading.
-                app = apps.app_configs[model._meta.app_label]
-                attrs['__module__'] = app.name      # full dotted name
+            # Abuse an internal API because the app registry is loading.
+            app = apps.app_configs[model._meta.app_label]
+            models_module = app.name
+            attrs['__module__'] = models_module
 
         fields = self.copy_fields(model)
         attrs.update(fields)
         attrs.update(self.get_extra_fields(model, fields))
         # type in python2 wants str as a first argument
         attrs.update(Meta=type(str('Meta'), (), self.get_meta_options(model)))
+        if self.table_name is not None:
+            attrs['Meta'].db_table = self.table_name
         name = 'Historical%s' % model._meta.object_name
         registered_models[model._meta.db_table] = model
         return python_2_unicode_compatible(
             type(str(name), self.bases, attrs))
+
+    def fields_included(self, model):
+        fields = []
+        for field in model._meta.fields:
+            if field.name not in self.excluded_fields:
+                fields.append(field)
+        return fields
 
     def copy_fields(self, model):
         """
@@ -128,30 +152,37 @@ class HistoricalRecords(object):
         a dictionary mapping field name to copied field object.
         """
         fields = {}
-        for field in model._meta.fields:
+        for field in self.fields_included(model):
             field = copy.copy(field)
-            field.rel = copy.copy(field.rel)
+            field.remote_field = copy.copy(field.remote_field)
             if isinstance(field, OrderWrt):
                 # OrderWrt is a proxy field, switch to a plain IntegerField
                 field.__class__ = models.IntegerField
             if isinstance(field, models.ForeignKey):
                 old_field = field
-                field_arguments = {}
-                if (getattr(old_field, 'one_to_one', False) or
-                        isinstance(old_field, models.OneToOneField)):
+                field_arguments = {'db_constraint': False}
+                if getattr(old_field, 'one_to_one', False) \
+                   or isinstance(old_field, models.OneToOneField):
                     FieldType = models.ForeignKey
                 else:
                     FieldType = type(old_field)
-                if django.get_version() >= "1.6":
-                    field_arguments['db_constraint'] = False
                 if getattr(old_field, 'to_fields', []):
                     field_arguments['to_field'] = old_field.to_fields[0]
-                elif django.get_version() < "1.6" and old_field.rel.field_name != 'id':
-                    field_arguments['to_field'] = old_field.rel.field_name
                 if getattr(old_field, 'db_column', None):
                     field_arguments['db_column'] = old_field.db_column
+
+                # If old_field.remote_field.model is 'self' then we have a
+                # case where object has a foreign key to itself. In this case
+                # we need to set the `model` value of the field to a model. We
+                # can use the old_field.model value.
+                if isinstance(old_field.remote_field.model, str) and \
+                   old_field.remote_field.model == 'self':
+                    object_to = old_field.model
+                else:
+                    object_to = old_field.remote_field.model
+
                 field = FieldType(
-                    old_field.rel.to,
+                    object_to,
                     related_name='+',
                     null=True,
                     blank=True,
@@ -173,36 +204,73 @@ class HistoricalRecords(object):
 
         user_model = getattr(settings, 'AUTH_USER_MODEL', 'auth.User')
 
-        @models.permalink
         def revert_url(self):
             """URL for this change in the default admin site."""
             opts = model._meta
-            try:
-                app_label, model_name = opts.app_label, opts.model_name
-            except AttributeError:  # Django < 1.7
-                app_label, model_name = opts.app_label, opts.module_name
-            return ('%s:%s_%s_simple_history' %
-                    (admin.site.name, app_label, model_name),
-                    [getattr(self, opts.pk.attname), self.history_id])
+            app_label, model_name = opts.app_label, opts.model_name
+            return reverse(
+                '%s:%s_%s_simple_history' % (
+                    admin.site.name,
+                    app_label,
+                    model_name
+                ),
+                args=[getattr(self, opts.pk.attname), self.history_id]
+            )
 
         def get_instance(self):
-            return model(**dict([(field.attname, getattr(self, field.attname))
-                                for field in fields.values()]))
+            return model(**{
+                field.attname: getattr(self, field.attname)
+                for field in fields.values()
+            })
+
+        def get_next_record(self):
+            """
+            Get the next history record for the instance. `None` if last.
+            """
+            return self.instance.history.filter(
+                Q(history_date__gt=self.history_date)
+            ).order_by('history_date').first()
+
+        def get_prev_record(self):
+            """
+            Get the previous history record for the instance. `None` if first.
+            """
+            return self.instance.history.filter(
+                Q(history_date__lt=self.history_date)
+            ).order_by('history_date').last()
+
+        if self.history_id_field:
+            history_id_field = self.history_id_field
+            history_id_field.primary_key = True
+            history_id_field.editable = False
+        elif getattr(settings, 'SIMPLE_HISTORY_HISTORY_ID_USE_UUID', False):
+            history_id_field = models.UUIDField(
+                primary_key=True, default=uuid.uuid4, editable=False
+            )
+        else:
+            history_id_field = models.AutoField(primary_key=True)
 
         return {
-            'history_id': models.AutoField(primary_key=True),
+            'history_id': history_id_field,
             'history_date': models.DateTimeField(),
+            'history_change_reason': models.CharField(max_length=100,
+                                                      null=True),
             'history_user': models.ForeignKey(
                 user_model, null=True, related_name=self.user_related_name,
                 on_delete=models.SET_NULL),
             'history_type': models.CharField(max_length=1, choices=(
-                ('+', 'Created'),
-                ('~', 'Changed'),
-                ('-', 'Deleted'),
+                ('+', _('Created')),
+                ('~', _('Changed')),
+                ('-', _('Deleted')),
             )),
-            'history_object': HistoricalObjectDescriptor(model),
+            'history_object': HistoricalObjectDescriptor(
+                model,
+                self.fields_included(model)
+            ),
             'instance': property(get_instance),
             'instance_type': model,
+            'next_record': property(get_next_record),
+            'prev_record': property(get_prev_record),
             'revert_url': revert_url,
             '__str__': lambda self: '%s as of %s' % (self.history_object,
                                                      self.history_date)
@@ -220,8 +288,8 @@ class HistoricalRecords(object):
         if self.user_set_verbose_name:
             name = self.user_set_verbose_name
         else:
-            name = string_concat('historical ',
-                                 smart_text(model._meta.verbose_name))
+            name = format_lazy('historical {}',
+                               smart_text(model._meta.verbose_name))
         meta_fields['verbose_name'] = name
         return meta_fields
 
@@ -237,12 +305,14 @@ class HistoricalRecords(object):
     def create_historical_record(self, instance, history_type):
         history_date = getattr(instance, '_history_date', now())
         history_user = self.get_history_user(instance)
+        history_change_reason = getattr(instance, 'changeReason', None)
         manager = getattr(instance, self.manager_name)
         attrs = {}
-        for field in instance._meta.fields:
+        for field in self.fields_included(instance):
             attrs[field.attname] = getattr(instance, field.attname)
         manager.create(history_date=history_date, history_type=history_type,
-                       history_user=history_user, **attrs)
+                       history_user=history_user,
+                       history_change_reason=history_change_reason, **attrs)
 
     def get_history_user(self, instance):
         """Get the modifying user from instance or middleware."""
@@ -250,86 +320,11 @@ class HistoricalRecords(object):
             return instance._history_user
         except AttributeError:
             try:
-                if self.thread.request.user.is_authenticated():
+                if self.thread.request.user.is_authenticated:
                     return self.thread.request.user
                 return None
             except AttributeError:
                 return None
-
-
-class CustomForeignKeyField(models.ForeignKey):
-
-    def __init__(self, *args, **kwargs):
-        warnings.warn("CustomForeignKeyField is deprecated.",
-                      DeprecationWarning)
-        super(CustomForeignKeyField, self).__init__(*args, **kwargs)
-        self.db_constraint = False
-        self.generate_reverse_relation = False
-
-    def get_attname(self):
-        return self.name
-
-    def get_one_to_one_field(self, to_field, other):
-        # HACK This creates a new custom foreign key based on to_field,
-        # and calls itself with that, effectively making the calls
-        # recursive
-        temp_field = self.__class__(to_field.rel.to._meta.object_name)
-        for key, val in to_field.__dict__.items():
-            if (isinstance(key, six.string_types)
-                    and not key.startswith('_')):
-                setattr(temp_field, key, val)
-        field = self.__class__.get_field(
-            temp_field, other, to_field.rel.to)
-        return field
-
-    def get_field(self, other, cls):
-        # this hooks into contribute_to_class() and this is
-        # called specifically after the class_prepared signal
-        to_field = copy.copy(self.rel.to._meta.pk)
-        field = self
-        if isinstance(to_field, models.OneToOneField):
-            field = self.get_one_to_one_field(to_field, other)
-        elif isinstance(to_field, models.AutoField):
-            field.__class__ = convert_auto_field(to_field)
-        else:
-            field.__class__ = to_field.__class__
-            excluded_prefixes = ("_", "__")
-            excluded_attributes = (
-                "rel",
-                "creation_counter",
-                "validators",
-                "error_messages",
-                "attname",
-                "column",
-                "help_text",
-                "name",
-                "model",
-                "unique_for_year",
-                "unique_for_date",
-                "unique_for_month",
-                "db_tablespace",
-                "db_index",
-                "db_column",
-                "default",
-                "auto_created",
-                "null",
-                "blank",
-            )
-            for key, val in to_field.__dict__.items():
-                if (isinstance(key, six.string_types)
-                        and not key.startswith(excluded_prefixes)
-                        and key not in excluded_attributes):
-                    setattr(field, key, val)
-        return field
-
-    def do_related_class(self, other, cls):
-        field = self.get_field(other, cls)
-        transform_field(field)
-        field.rel = None
-
-    def contribute_to_class(self, cls, name):
-        # HACK: remove annoying descriptor (don't super())
-        RelatedField.contribute_to_class(self, cls, name)
 
 
 def transform_field(field):
@@ -362,17 +357,19 @@ def convert_auto_field(field):
     must be replaced with an IntegerField.
     """
     connection = router.db_for_write(field.model)
-    if settings.DATABASES[connection]['ENGINE'] in ('django_mongodb_engine',):
+    if settings.DATABASES[connection].get('ENGINE') in \
+       ('django_mongodb_engine',):
         # Check if AutoField is string for django-non-rel support
         return models.TextField
     return models.IntegerField
 
 
 class HistoricalObjectDescriptor(object):
-    def __init__(self, model):
+    def __init__(self, model, fields_included):
         self.model = model
+        self.fields_included = fields_included
 
     def __get__(self, instance, owner):
         values = (getattr(instance, f.attname)
-                  for f in self.model._meta.fields)
+                  for f in self.fields_included)
         return self.model(*values)