+++ simple_history/tests/admin.py
@@ -3,7 +3,16 @@ from __future__ import unicode_literals
 from django.contrib import admin
 
 from simple_history.admin import SimpleHistoryAdmin
-from .models import Poll, Choice, Person, Book, Document, Paper, Employee
+from .models import (
+    Book,
+    Choice,
+    ConcreteExternal,
+    Document,
+    Employee,
+    Paper,
+    Person,
+    Poll
+)
 
 
 class PersonAdmin(SimpleHistoryAdmin):
@@ -11,10 +20,15 @@ class PersonAdmin(SimpleHistoryAdmin):
         return False
 
 
+class ChoiceAdmin(SimpleHistoryAdmin):
+    history_list_display = ['votes']
+
+
 admin.site.register(Poll, SimpleHistoryAdmin)
-admin.site.register(Choice, SimpleHistoryAdmin)
+admin.site.register(Choice, ChoiceAdmin)
 admin.site.register(Person, PersonAdmin)
 admin.site.register(Book, SimpleHistoryAdmin)
 admin.site.register(Document, SimpleHistoryAdmin)
 admin.site.register(Paper, SimpleHistoryAdmin)
 admin.site.register(Employee, SimpleHistoryAdmin)
+admin.site.register(ConcreteExternal, SimpleHistoryAdmin)