+++ simple_history/tests/tests/test_models.py
@@ -1,38 +1,63 @@
 from __future__ import unicode_literals
 
-from datetime import datetime, timedelta
 import unittest
+import uuid
+import warnings
+from datetime import datetime, timedelta
 
-import django
+from django.apps import apps
+from django.contrib.auth import get_user_model
+from django.core.files.base import ContentFile
 from django.db import models
-from django.db.models.loading import get_model
 from django.db.models.fields.proxy import OrderWrt
 from django.test import TestCase
-from django.core.files.base import ContentFile
 
 from simple_history.models import HistoricalRecords, convert_auto_field
-from simple_history import register
-from ..models import (
-    AdminProfile, Bookcase, MultiOneToOne, Poll, Choice, Voter, Restaurant,
-    Person, FileModel, Document, Book, HistoricalPoll, Library, State,
-    AbstractBase, ConcreteAttr, ConcreteUtil, SelfFK, Temperature, WaterLevel,
-    ExternalModel1, ExternalModel3, UnicodeVerboseName, HistoricalChoice,
-    HistoricalState, HistoricalCustomFKError, Series, SeriesWork, PollInfo,
-    UserAccessorDefault, UserAccessorOverride, Employee, Country, Province,
-    City
-)
+from simple_history.utils import update_change_reason
 from ..external.models import ExternalModel2, ExternalModel4
+from ..models import (
+    AbstractBase,
+    AdminProfile,
+    Book,
+    Bookcase,
+    Choice,
+    City,
+    ConcreteAttr,
+    ConcreteExternal,
+    ConcreteUtil,
+    Contact,
+    ContactRegister,
+    Country,
+    Document,
+    Employee,
+    ExternalModel1,
+    ExternalModel3,
+    FileModel,
+    HistoricalChoice,
+    HistoricalCustomFKError,
+    HistoricalPoll,
+    HistoricalState,
+    Library,
+    MultiOneToOne,
+    Person,
+    Poll,
+    PollInfo,
+    PollWithExcludeFields,
+    Province,
+    Restaurant,
+    SelfFK,
+    Series,
+    SeriesWork,
+    State,
+    Temperature,
+    UnicodeVerboseName,
+    UUIDModel,
+    UUIDDefaultModel,
+    WaterLevel
+)
 
-try:
-    from unittest import skipUnless
-except ImportError:
-    from unittest2 import skipUnless
-try:
-    from django.contrib.auth import get_user_model
-    User = get_user_model()
-except ImportError:  # django 1.4 compatibility
-    from django.contrib.auth.models import User
-
+get_model = apps.get_model
+User = get_user_model()
 today = datetime(2021, 1, 1, 10, 0)
 tomorrow = today + timedelta(days=1)
 yesterday = today - timedelta(days=1)
@@ -54,7 +79,7 @@ class HistoricalRecordsTest(TestCase):
             self.assertEqual(getattr(record, key), value)
         self.assertEqual(record.history_object.__class__, klass)
         for key, value in values_dict.items():
-            if key != 'history_type':
+            if key not in ['history_type', 'history_change_reason']:
                 self.assertEqual(getattr(record.history_object, key), value)
 
     def test_create(self):
@@ -74,36 +99,63 @@ class HistoricalRecordsTest(TestCase):
         p = Poll.objects.get()
         p.pub_date = tomorrow
         p.save()
+        update_change_reason(p, 'future poll')
         update_record, create_record = p.history.all()
         self.assertRecordValues(create_record, Poll, {
             'question': "what's up?",
             'pub_date': today,
             'id': p.id,
+            'history_change_reason': None,
             'history_type': "+"
         })
         self.assertRecordValues(update_record, Poll, {
             'question': "what's up?",
             'pub_date': tomorrow,
             'id': p.id,
+            'history_change_reason': 'future poll',
             'history_type': "~"
         })
         self.assertDatetimesEqual(update_record.history_date, datetime.now())
 
-    def test_delete(self):
+    def test_delete_verify_change_reason_implicitly(self):
         p = Poll.objects.create(question="what's up?", pub_date=today)
         poll_id = p.id
+        p.changeReason = 'wrongEntry'
         p.delete()
         delete_record, create_record = Poll.history.all()
         self.assertRecordValues(create_record, Poll, {
             'question': "what's up?",
             'pub_date': today,
             'id': poll_id,
+            'history_change_reason': None,
             'history_type': "+"
         })
         self.assertRecordValues(delete_record, Poll, {
             'question': "what's up?",
             'pub_date': today,
             'id': poll_id,
+            'history_change_reason': 'wrongEntry',
+            'history_type': "-"
+        })
+
+    def test_delete_verify_change_reason_explicity(self):
+        p = Poll.objects.create(question="what's up?", pub_date=today)
+        poll_id = p.id
+        p.delete()
+        update_change_reason(p, 'wrongEntry')
+        delete_record, create_record = Poll.history.all()
+        self.assertRecordValues(create_record, Poll, {
+            'question': "what's up?",
+            'pub_date': today,
+            'id': poll_id,
+            'history_change_reason': None,
+            'history_type': "+"
+        })
+        self.assertRecordValues(delete_record, Poll, {
+            'question': "what's up?",
+            'pub_date': today,
+            'id': poll_id,
+            'history_change_reason': 'wrongEntry',
             'history_type': "-"
         })
 
@@ -164,7 +216,8 @@ class HistoricalRecordsTest(TestCase):
         lib = Library.objects.create()
         state = State.objects.create(library=lib)
         self.assertTrue(hasattr(lib, 'state_set'))
-        self.assertIsNone(state._meta.get_field('library').rel.related_name,
+        self.assertIsNone(state._meta.get_field('library')
+                          .remote_field.related_name,
                           "the '+' shouldn't leak through to the original "
                           "model's field related_name")
 
@@ -271,7 +324,8 @@ class HistoricalRecordsTest(TestCase):
         country = Country.objects.create(code='US')
         city = City.objects.create(country=country)
         country_field = City._meta.get_field('country')
-        self.assertTrue(getattr(country_field, 'db_column') in str(city.history.all().query))
+        self.assertIn(getattr(country_field, 'db_column'),
+                      str(city.history.all().query))
 
     def test_raw_save(self):
         document = Document()
@@ -296,10 +350,10 @@ class HistoricalRecordsTest(TestCase):
         self.assertEqual('dead trees', b.history.all()[0]._meta.verbose_name)
 
     def test_historical_verbose_name_follows_model_verbose_name(self):
-        l = Library()
-        l.save()
+        library = Library()
+        library.save()
         self.assertEqual('historical quiet please',
-                         l.history.get()._meta.verbose_name)
+                         library.history.get()._meta.verbose_name)
 
     def test_foreignkey_primarykey(self):
         """Test saving a tracked model with a `ForeignKey` primary key."""
@@ -308,39 +362,99 @@ class HistoricalRecordsTest(TestCase):
         poll_info = PollInfo(poll=poll)
         poll_info.save()
 
+    def test_model_with_excluded_fields(self):
+        p = PollWithExcludeFields(question="what's up?", pub_date=today)
+        p.save()
+        history = PollWithExcludeFields.history.all()[0]
+        all_fields_names = [f.name for f in history._meta.fields]
+        self.assertIn('question', all_fields_names)
+        self.assertNotIn('pub_date', all_fields_names)
 
-class RegisterTest(TestCase):
-    def test_register_no_args(self):
-        self.assertEqual(len(Choice.history.all()), 0)
-        poll = Poll.objects.create(pub_date=today)
-        choice = Choice.objects.create(poll=poll, votes=0)
-        self.assertEqual(len(choice.history.all()), 1)
+    def test_uuid_history_id(self):
+        entry = UUIDModel.objects.create()
 
-    def test_register_separate_app(self):
-        get_history = lambda model: model.history
-        self.assertRaises(AttributeError, get_history, User)
-        self.assertEqual(len(User.histories.all()), 0)
-        user = User.objects.create(username='bob', password='pass')
-        self.assertEqual(len(User.histories.all()), 1)
-        self.assertEqual(len(user.histories.all()), 1)
+        history = entry.history.all()[0]
+        self.assertTrue(isinstance(history.history_id, uuid.UUID))
 
-    def test_reregister(self):
-        register(Restaurant, manager_name='again')
-        register(User, manager_name='again')
-        self.assertTrue(hasattr(Restaurant, 'updates'))
-        self.assertFalse(hasattr(Restaurant, 'again'))
-        self.assertTrue(hasattr(User, 'histories'))
-        self.assertFalse(hasattr(User, 'again'))
+    def test_uuid_default_history_id(self):
+        entry = UUIDDefaultModel.objects.create()
 
-    def test_register_custome_records(self):
-        self.assertEqual(len(Voter.history.all()), 0)
-        poll = Poll.objects.create(pub_date=today)
-        choice = Choice.objects.create(poll=poll, votes=0)
-        user = User.objects.create(username='voter')
-        voter = Voter.objects.create(choice=choice, user=user)
-        self.assertEqual(len(voter.history.all()), 1)
-        expected = 'Voter object changed by None as of '
-        self.assertEqual(expected, str(voter.history.all()[0])[:len(expected)])
+        history = entry.history.all()[0]
+        self.assertTrue(isinstance(history.history_id, uuid.UUID))
+
+    def test_get_prev_record(self):
+        poll = Poll(question="what's up?", pub_date=today)
+        poll.save()
+        poll.question = "ask questions?"
+        poll.save()
+        poll.question = "eh?"
+        poll.save()
+        poll.question = "one more?"
+        poll.save()
+        first_record = poll.history.filter(question="what's up?").get()
+        second_record = poll.history.filter(question="ask questions?").get()
+        third_record = poll.history.filter(question="eh?").get()
+        fourth_record = poll.history.filter(question="one more?").get()
+        self.assertIsNone(first_record.prev_record)
+
+        def assertRecordsMatch(record_a, record_b):
+            self.assertEqual(record_a, record_b)
+            self.assertEqual(record_a.question, record_b.question)
+        assertRecordsMatch(second_record.prev_record, first_record)
+        assertRecordsMatch(third_record.prev_record, second_record)
+        assertRecordsMatch(fourth_record.prev_record, third_record)
+
+    def test_get_prev_record_none_if_only(self):
+        poll = Poll(question="what's up?", pub_date=today)
+        poll.save()
+        self.assertEqual(poll.history.count(), 1)
+        record = poll.history.get()
+        self.assertIsNone(record.prev_record)
+
+    def test_get_prev_record_none_if_earliest(self):
+        poll = Poll(question="what's up?", pub_date=today)
+        poll.save()
+        poll.question = "ask questions?"
+        poll.save()
+        first_record = poll.history.filter(question="what's up?").get()
+        self.assertIsNone(first_record.prev_record)
+
+    def test_get_next_record(self):
+        poll = Poll(question="what's up?", pub_date=today)
+        poll.save()
+        poll.question = "ask questions?"
+        poll.save()
+        poll.question = "eh?"
+        poll.save()
+        poll.question = "one more?"
+        poll.save()
+        first_record = poll.history.filter(question="what's up?").get()
+        second_record = poll.history.filter(question="ask questions?").get()
+        third_record = poll.history.filter(question="eh?").get()
+        fourth_record = poll.history.filter(question="one more?").get()
+        self.assertIsNone(fourth_record.next_record)
+
+        def assertRecordsMatch(record_a, record_b):
+            self.assertEqual(record_a, record_b)
+            self.assertEqual(record_a.question, record_b.question)
+        assertRecordsMatch(first_record.next_record, second_record)
+        assertRecordsMatch(second_record.next_record, third_record)
+        assertRecordsMatch(third_record.next_record, fourth_record)
+
+    def test_get_next_record_none_if_only(self):
+        poll = Poll(question="what's up?", pub_date=today)
+        poll.save()
+        self.assertEqual(poll.history.count(), 1)
+        record = poll.history.get()
+        self.assertIsNone(record.next_record)
+
+    def test_get_next_record_none_if_most_recent(self):
+        poll = Poll(question="what's up?", pub_date=today)
+        poll.save()
+        poll.question = "ask questions?"
+        poll.save()
+        recent_record = poll.history.filter(question="ask questions?").get()
+        self.assertIsNone(recent_record.next_record)
 
 
 class CreateHistoryModelTests(unittest.TestCase):
@@ -349,8 +463,8 @@ class CreateHistoryModelTests(unittest.TestCase):
         records = HistoricalRecords()
         records.module = AdminProfile.__module__
         try:
-            records.create_history_model(AdminProfile)
-        except:
+            records.create_history_model(AdminProfile, False)
+        except Exception:
             self.fail("SimpleHistory should handle foreign keys to one to one"
                       "fields to integer fields without throwing an exception")
 
@@ -358,8 +472,8 @@ class CreateHistoryModelTests(unittest.TestCase):
         records = HistoricalRecords()
         records.module = Bookcase.__module__
         try:
-            records.create_history_model(Bookcase)
-        except:
+            records.create_history_model(Bookcase, False)
+        except Exception:
             self.fail("SimpleHistory should handle foreign keys to one to one"
                       "fields to char fields without throwing an exception.")
 
@@ -367,8 +481,8 @@ class CreateHistoryModelTests(unittest.TestCase):
         records = HistoricalRecords()
         records.module = MultiOneToOne.__module__
         try:
-            records.create_history_model(MultiOneToOne)
-        except:
+            records.create_history_model(MultiOneToOne, False)
+        except Exception:
             self.fail("SimpleHistory should handle foreign keys to one to one"
                       "fields to one to one fields without throwing an "
                       "exception.")
@@ -400,6 +514,10 @@ class AppLabelTest(TestCase):
                          'external_externalmodel4')
         self.assertEqual(self.get_table_name(ExternalModel4.histories),
                          'tests_historicalexternalmodel4')
+        self.assertEqual(self.get_table_name(ConcreteExternal.objects),
+                         'tests_concreteexternal')
+        self.assertEqual(self.get_table_name(ConcreteExternal.history),
+                         'tests_historicalconcreteexternal')
 
     def test_get_model(self):
         self.assertEqual(get_model('external', 'ExternalModel1'),
@@ -421,6 +539,13 @@ class AppLabelTest(TestCase):
                          ExternalModel4)
         self.assertEqual(get_model('tests', 'HistoricalExternalModel4'),
                          ExternalModel4.histories.model)
+
+        # Test that historical model is defined within app of concrete
+        # model rather than abstract base model
+        self.assertEqual(get_model('tests', 'ConcreteExternal'),
+                         ConcreteExternal)
+        self.assertEqual(get_model('tests', 'HistoricalConcreteExternal'),
+                         ConcreteExternal.history.model)
 
 
 class HistoryManagerTest(TestCase):
@@ -464,7 +589,10 @@ class HistoryManagerTest(TestCase):
         most_recent = poll.history.most_recent()
         self.assertEqual(most_recent.question, "why?")
         times = [r.history_date for r in poll.history.all()]
-        question_as_of = lambda time: poll.history.as_of(time).question
+
+        def question_as_of(time):
+            return poll.history.as_of(time).question
+
         self.assertEqual(question_as_of(times[0]), "why?")
         self.assertEqual(question_as_of(times[1]), "how's it going?")
         self.assertEqual(question_as_of(times[2]), "what's up?")
@@ -488,7 +616,10 @@ class HistoryManagerTest(TestCase):
         most_recent = choice.history.most_recent()
         self.assertEqual(most_recent.poll.pk, how_poll.pk)
         times = [r.history_date for r in choice.history.all()]
-        poll_as_of = lambda time: choice.history.as_of(time).poll
+
+        def poll_as_of(time):
+            return choice.history.as_of(time).poll
+
         self.assertEqual(poll_as_of(times[0]).pk, how_poll.pk)
         self.assertEqual(poll_as_of(times[1]).pk, why_poll.pk)
 
@@ -506,22 +637,15 @@ class HistoryManagerTest(TestCase):
             self.assertRaises(TypeError, HistoricalRecords, bases=bases)
 
     def test_import_related(self):
-        field_object = HistoricalChoice._meta.get_field_by_name('poll')[0]
-        try:
-            related_model = field_object.rel.related_model
-        except AttributeError:  # Django<1.8
-            related_model = field_object.related.model
+        field_object = HistoricalChoice._meta.get_field('poll')
+        related_model = field_object.remote_field.related_model
         self.assertEqual(related_model, HistoricalChoice)
 
     def test_string_related(self):
-        field_object = HistoricalState._meta.get_field_by_name('library')[0]
-        try:
-            related_model = field_object.rel.related_model
-        except AttributeError:  # Django<1.8
-            related_model = field_object.related.model
+        field_object = HistoricalState._meta.get_field('library')
+        related_model = field_object.remote_field.related_model
         self.assertEqual(related_model, HistoricalState)
 
-    @skipUnless(django.get_version() >= "1.7", "Requires 1.7 migrations")
     def test_state_serialization_of_customfk(self):
         from django.db.migrations import state
         state.ModelState.from_model(HistoricalCustomFKError)
@@ -541,9 +665,13 @@ class TestConvertAutoField(TestCase):
 
         Default Django ORM uses an integer-based auto field.
         """
-        with self.settings(DATABASES={'default': {
-                'ENGINE': 'django.db.backends.postgresql_psycopg2'}}):
-            assert convert_auto_field(self.field) == models.IntegerField
+        with warnings.catch_warnings():
+            warnings.filterwarnings('ignore', category=UserWarning,
+                                    message='Overriding setting DATABASES ' +
+                                            'can lead to unexpected behavior.')
+            with self.settings(DATABASES={'default': {
+                    'ENGINE': 'django.db.backends.postgresql_psycopg2'}}):
+                assert convert_auto_field(self.field) == models.IntegerField
 
     def test_non_relational(self):
         """Non-relational test
@@ -551,9 +679,14 @@ class TestConvertAutoField(TestCase):
         MongoDB uses a string-based auto field. We need to make sure
         the converted field type is string.
         """
-        with self.settings(DATABASES={'default': {
-                'ENGINE': 'django_mongodb_engine'}}):
-            assert convert_auto_field(self.field) == models.TextField
+
+        with warnings.catch_warnings():
+            warnings.filterwarnings('ignore', category=UserWarning,
+                                    message='Overriding setting DATABASES ' +
+                                            'can lead to unexpected behavior.')
+            with self.settings(DATABASES={'default': {
+                    'ENGINE': 'django_mongodb_engine'}}):
+                assert convert_auto_field(self.field) == models.TextField
 
 
 class TestOrderWrtField(TestCase):
@@ -593,7 +726,7 @@ class TestOrderWrtField(TestCase):
             self.w_horse.pk,
             self.w_nephew.pk,
             self.w_battle.pk]
-        self.assertEqual(order, expected)
+        self.assertSequenceEqual(order, expected)
         self.assertEqual(0, self.w_lion._order)
         self.assertEqual(1, self.w_caspian._order)
         self.assertEqual(2, self.w_voyage._order)
@@ -603,11 +736,11 @@ class TestOrderWrtField(TestCase):
         self.assertEqual(6, self.w_battle._order)
 
     def test_order_field_in_historical_model(self):
-        work_order_field = self.w_lion._meta.get_field_by_name('_order')[0]
+        work_order_field = self.w_lion._meta.get_field('_order')
         self.assertEqual(type(work_order_field), OrderWrt)
 
         history = self.w_lion.history.all()[0]
-        history_order_field = history._meta.get_field_by_name('_order')[0]
+        history_order_field = history._meta.get_field('_order')
         self.assertEqual(type(history_order_field), models.IntegerField)
 
     def test_history_object_has_order(self):
@@ -631,7 +764,8 @@ class TestOrderWrtField(TestCase):
             self.w_chair.pk,
             self.w_battle.pk]
         self.series.set_serieswork_order(chronological)
-        self.assertEqual(self.series.get_serieswork_order(), chronological)
+        self.assertSequenceEqual(self.series.get_serieswork_order(),
+                                 chronological)
 
         # This uses an update, not a save, so no new history is created
         w_caspian = SeriesWork.objects.get(id=self.w_caspian.id)
@@ -663,7 +797,6 @@ class TestOrderWrtField(TestCase):
         self.assertEqual(order[5], self.w_chair.pk)
         self.assertEqual(order[6], self.w_battle.pk)
 
-    @skipUnless(django.get_version() >= "1.7", "Requires 1.7 migrations")
     def test_migrations_include_order(self):
         from django.db.migrations import state
         model_state = state.ModelState.from_model(SeriesWork.history.model)
@@ -713,17 +846,6 @@ class TestLatest(TestCase):
         assert HistoricalPoll.objects.latest().pk == 1
 
 
-class TestUserAccessor(unittest.TestCase):
-
-    def test_accessor_default(self):
-        register(UserAccessorDefault)
-        assert not hasattr(User, 'historicaluseraccessordefault_set')
-
-    def test_accessor_override(self):
-        register(UserAccessorOverride, user_related_name='my_history_model_accessor')
-        assert hasattr(User, 'my_history_model_accessor')
-
-
 class TestMissingOneToOne(TestCase):
 
     def setUp(self):
@@ -745,3 +867,22 @@ class TestMissingOneToOne(TestCase):
         self.assertEqual(original.manager_id, 1)
         with self.assertRaises(Employee.DoesNotExist):
             original.manager
+
+
+class CustomTableNameTest1(TestCase):
+
+    @staticmethod
+    def get_table_name(manager):
+        return manager.model._meta.db_table
+
+    def test_custom_table_name(self):
+        self.assertEqual(
+            self.get_table_name(Contact.history),
+            'contacts_history',
+        )
+
+    def test_custom_table_name_from_register(self):
+        self.assertEqual(
+            self.get_table_name(ContactRegister.history),
+            'contacts_register_history',
+        )